Detect/requirements.txt
alabaster==0.7.12
arm-pyart==1.10.1
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
botocore==1.12.183
certifi==2019.6.16
cffi==1.12.3
cftime==1.0.3.4
chardet==3.0.4
cloudpickle==1.1.1
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dispy==4.11.0
docutils==0.14
entrypoints==0.3
idna==2.8
imagesize==1.1.0
intel-openmp==2019.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mkl==2019.0
nbconvert==5.5.0
nbformat==4.4.0
netCDF4==1.4.2
numpy==1.15.4
numpydoc==0.9.1
opencv-python==4.1.0.25
packaging==19.0
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==2.0.9
psutil==5.6.2
pycodestyle==2.5.0
pycos==4.8.11
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
PyQtWebEngine==5.12.1
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
pywin32==223
pywin32-ctypes==0.2.0
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.7.1
requests==2.22.0
rope==0.14.0
s3transfer==0.2.1
scipy==1.2.1
sip==4.19.8
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.8
Sphinx==2.1.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder-kernels==0.4.4
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.1
